+++ naiveproxy/src/net/url_request/url_request_http_job.cc
@@ -455,6 +455,7 @@ void URLRequestHttpJob::Start() {
 #if BUILDFLAG(ENABLE_REPORTING)
   request_info_.reporting_upload_depth = request_->reporting_upload_depth();
 #endif
+  request_info_.is_shared_resource = request_->is_shared_resource();
 
   CookieStore* cookie_store = request()->context()->cookie_store();
   const CookieAccessDelegate* delegate =
@@ -701,7 +702,7 @@ void URLRequestHttpJob::StartTransactionInternal() {
   // If we already have a transaction, then we should restart the transaction
   // with auth provided by auth_credentials_.
 
-  int rv;
+  int rv = OK;
 
   // Notify NetworkQualityEstimator.
   NetworkQualityEstimator* network_quality_estimator =
@@ -716,11 +717,12 @@ void URLRequestHttpJob::StartTransactionInternal() {
     auth_credentials_ = AuthCredentials();
   } else {
     DCHECK(request_->context()->http_transaction_factory());
+    transaction_ =
+        request_->context()->http_transaction_factory()->CreateTransaction(
+            priority_);
+    CHECK(transaction_);
 
-    rv = request_->context()->http_transaction_factory()->CreateTransaction(
-        priority_, &transaction_);
-
-    if (rv == OK && request_info_.url.SchemeIsWSOrWSS()) {
+    if (request_info_.url.SchemeIsWSOrWSS()) {
       base::SupportsUserData::Data* data =
           request_->GetUserData(kWebSocketHandshakeUserDataKey);
       if (data) {
@@ -1171,29 +1173,29 @@ void URLRequestHttpJob::OnSetCookieResult(const CookieOptions& options,
 
 #if BUILDFLAG(ENABLE_DEVICE_BOUND_SESSIONS)
 void URLRequestHttpJob::ProcessDeviceBoundSessionsHeader() {
-  if (!request_->allows_device_bound_session_registration() &&
-      !features::kDeviceBoundSessionsForceEnableForTesting.Get()) {
-    return;
-  }
-
   device_bound_sessions::SessionService* service =
       request_->context()->device_bound_session_service();
   if (!service) {
     return;
   }
 
+  const auto& request_url = request_->url();
+  auto* headers = GetResponseHeaders();
+
   // If response header Sec-Session-Registration is present and configured
   // appropriately, trigger a registration request per header value to attempt
   // to create a new session.
-  const auto& request_url = request_->url();
-  auto* headers = GetResponseHeaders();
-  std::vector<device_bound_sessions::RegistrationFetcherParam> params =
-      device_bound_sessions::RegistrationFetcherParam::CreateIfValid(
-          request_url, headers);
-  for (auto& param : params) {
-    service->RegisterBoundSession(
-        request_->device_bound_session_access_callback(), std::move(param),
-        request_->isolation_info(), request_->net_log(), request_->initiator());
+  if (request_->allows_device_bound_session_registration() ||
+      features::kDeviceBoundSessionsForceEnableForTesting.Get()) {
+    std::vector<device_bound_sessions::RegistrationFetcherParam> params =
+        device_bound_sessions::RegistrationFetcherParam::CreateIfValid(
+            request_url, headers);
+    for (auto& param : params) {
+      service->RegisterBoundSession(
+          request_->device_bound_session_access_callback(), std::move(param),
+          request_->isolation_info(), request_->net_log(),
+          request_->initiator());
+    }
   }
 
   // If response header Sec-Session-Challenge is present and configured